card数据查询

+++ packgopay/src/main/java/com/pakgopay/common/response/report/BaseReportResponse.java
@@ -27,5 +27,5 @@ public class BaseReportResponse {
     /**
      * total number
      */
-    private Long totalNumber;
+    private Integer totalNumber;
 }
