+++ packgopay/src/main/java/com/pakgopay/mapper/MerchantReportMapper.java
@@ -1,6 +1,6 @@
 package com.pakgopay.mapper;
 
-import com.pakgopay.entity.MerchantReportEntity;
+import com.pakgopay.entity.report.MerchantReportEntity;
 import com.pakgopay.mapper.dto.MerchantReportDto;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -12,11 +12,11 @@ public interface MerchantReportMapper {
     /** Insert report */
     int insert(MerchantReportDto dto);
 
-    /** Update report (by userId + recordDate + orderType + currencyType) */
+    /** Update report (by userId + recordDate + orderType + currency) */
     int update(MerchantReportDto dto);
 
     /** Total count */
-    long countByQuery(MerchantReportEntity query);
+    Integer countByQuery(MerchantReportEntity query);
 
     /** Page query */
     List<MerchantReportDto> pageByQuery(MerchantReportEntity query);
